--- a/perf-regression/Makefile.subdir
+++ b/perf-regression/Makefile.subdir
@@ -1,5 +1,9 @@
 if HAVE_MARGO
-bin_PROGRAMS += perf-regression/margo-p2p-latency perf-regression/margo-p2p-bw perf-regression/margo-p2p-vector perf-regression/abt-eventual-bench
+bin_PROGRAMS += perf-regression/margo-p2p-latency \
+		perf-regression/margo-p2p-bw \
+		perf-regression/margo-p2p-vector \
+		perf-regression/abt-eventual-bench\
+		perf-regression/abt-future-bench
 endif
 
 if HAVE_SSG