--- a/tests/plot-abt-eventual-bench.sh
+++ b/tests/plot-abt-eventual-bench.sh
@@ -11,5 +11,6 @@ DAT=$1
 # itself, which is where the gnuplot scripts are that we want to use
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
-gnuplot -c $SCRIPT_DIR/rate.plt $DAT $DAT.rate.eog
+gnuplot -c $SCRIPT_DIR/rate.plt $DAT $DAT.rate.png
+gnuplot -c $SCRIPT_DIR/cpu.plt $DAT $DAT.cpu.png