--- a/tests/rate.plt
+++ b/tests/rate.plt
@@ -1,0 +1,17 @@
+#set data style lines
+#set style data linespoints
+set xlabel 'Number of ULTs'
+set ylabel 'Eventual rate (ops/s)'
+set term png medium
+#set term po eps mono dashed "Helvetica" 16
+#set term post eps color "Times-Roman" 20
+#set size 1.4,1.4
+#set key right bottom
+#set pointsize 2
+set key off
+
+INFILE=ARG1
+OUTFILE=ARG2
+
+set output OUTFILE
+plot INFILE using 1:4 with lines