This adds a benchmark called `abt-eventual-benchmark` that measures how well we can create/wait/set/destroy concurrent eventuals.  It approximates a concurrent mochi provider workload by creating N concurrent threads, each of which (in a loop) creates detached sub-threads that `margo_thread_sleep()` for a specified amount of time before setting an eventual and exiting.  The sleep is a stand-in for waiting on network activity.

Example, with a duration of 5 seconds, a sleep interval of 1 ms, and 16 ULTs:
```
> ./abt-eventual-bench -d 5 -i 1 -n 16
#<num_ults>	<test_duration_sec>	<interval_msec>	<ops/s>	<cpu_time_sec>
16	5	1	15952.800000	4.255606
```
The above example shows 16 ULTs getting nearly 16,000 operations per second, which is pretty much ideal (each can do 1000 per second if everything is working right because there is a 1ms delay in each operation).  The last column shows CPU time.  In this case a 5 second run took 4.2 seconds of user-space CPU time, which is not expected.

These results are most interesting if you sweep across a range of ULT counts to see how behavior changes with scale.  There are scripts included in this PR to do this and plot the results for an example scenario.  The scripts are in the tests/ subdirectory.  The first takes the eventual executable as it's argument (so that it can account for different build directories).  The second takes the .dat file produced by the first script as it's argument and produces two plots.
```
> tests/sweep-abt-eventual-bench.sh build/perf-regression/abt-eventual-bench
writing results to sweep-abt-eventual-bench.1122789.dat
... running benchmark with 100 ULTs
... running benchmark with 200 ULTs
... running benchmark with 300 ULTs
... running benchmark with 400 ULTs
... running benchmark with 500 ULTs
... running benchmark with 600 ULTs
... running benchmark with 700 ULTs
... running benchmark with 800 ULTs
... running benchmark with 900 ULTs
... running benchmark with 1000 ULTs
results are in sweep-abt-eventual-bench.1122789.dat
> tests/plot-abt-eventual-bench.sh sweep-abt-eventual-bench.1122789.dat
> ls *.png
sweep-abt-eventual-bench.1122789.dat.cpu.png
sweep-abt-eventual-bench.1122789.dat.rate.png
```
Example plots, showing high cpu usage in each configuration.   The eventual rate climbs linearly with the number of ULTs (as expected) until 500 concurrent ULTs, at which point throughput starts dropping rather climbing or plateauing.

![sweep-abt-eventual-bench 1122789 dat cpu](https://user-images.githubusercontent.com/5541937/153058728-3134f873-0faf-45b5-9647-10ce4be2ce1f.png)
![sweep-abt-eventual-bench 1122789 dat rate](https://user-images.githubusercontent.com/5541937/153058731-5c967c2e-2de2-4d1f-af32-952007035e01.png)

The above example is from a laptop with a debugging build; we need to repeat this on real systems.

This benchmark only uses one execution stream (the primary one) so throughput will be limited eventually by how many concurrent ULTs it can keep busy.

Before merging this we need to make sure the benchmark is working correctly, clean up the code a little, and capture instructions in a README.