@roblatham00 2022-02-08
On summit with argobots-1.1, running the test with `jsrun -r1 -n1 -a1`

![sweep-abt-eventual-bench 1966961 dat cpu](https://user-images.githubusercontent.com/2278417/153062798-66d26b46-8cc5-40e2-98f1-929b97b82b49.png)
![sweep-abt-eventual-bench 1966961 dat rate](https://user-images.githubusercontent.com/2278417/153062800-6c5aec04-c2b8-47af-8b8b-9174d3c42f49.png)