@carns 2022-02-08
Hunh.  So why the high CPU time, I wonder?  I'm not sure if I structured something wrong in the benchmark or if something is spinning that isn't supposed to.

Glad to see that performance is plateauing rather than dropping after hitting the peak at least.