@carns 2022-02-08
I wondered if the CPU usage was coming from the thread joins in the main routine (since we don't normally use that particular primitive in a Mochi server; the threads are detached).  I tested that by adding a margo_thread_sleep() in the main() routine before entering the join loop, though, and it didn't make a difference in CPU consumption.