@roblatham00 2022-02-08
crusher is of course bizzare
![sweep-abt-eventual-bench 27840 dat cpu](https://user-images.githubusercontent.com/2278417/153074442-4ca503da-a681-4354-a6c8-affcd7b7efcc.png)
![sweep-abt-eventual-bench 27840 dat rate](https://user-images.githubusercontent.com/2278417/153074447-9b5eea6b-7780-4158-bd72-f42fd1f64c3f.png)