@carns 2022-02-08
The crusher graph (the rate, not the cpu usage) is what we would hope for, I think.  It's should scale linearly in a perfect world.