@carns 2022-02-08
Maybe the busy spinning is just an artifact of the Mercury progress loop in Margo being cautious about the possibility of other ULTs needing to execute.  Can test that by initializing margo with a progress thread (in that case the loop should be allowed to sleep).