@carns 2022-02-08
Activating a progress thread does fix the cpu issue, but also seems to break the timers; there is another (different) problem to sort out in that mode it looks like.