@roblatham00 2022-02-09
I'm going to take a crack at a `futures` version