@carns 2022-10-04
Ok with me.  I think we stalled because some of the results were confusing, but I think we should just go ahead and merge it.  Some of the summit issues are likely fixed by changes that have happened in argobots@main since we were running these tests.

I'll let you do the merge @roblatham00 if you are comfortable with how the benchmarks look; I haven't tried running this in a while.