@roblatham00 2022-10-10
looking back at february's discussion: 
> Before merging this we need to make sure the benchmark is working correctly, clean up the code a little, and capture instructions in a README.

- [ ] ensure benchmark working correctly
- [ ] clean up code
- [ ] capture instructions in a README