@carns 2022-10-10
> Re-rrunning the experiment with latest code on summit
> 
> ```
> $ spack find --loaded | less
> -- linux-rhel8-power9le / gcc@11.1.0 ----------------------------
> argobots@main
> autoconf@2.69
> cereal@1.3.2
> cmake@3.20.2
> fmt@8.1.1
> gnuconfig@2021-08-14
> json-c@0.15
> libfabric@1.15.1
> m4@1.4.18
> mercury@master
> mochi-bedrock@develop
> mochi-margo@develop
> mochi-ssg@develop
> mochi-thallium@develop
> nlohmann-json@3.10.4
> pkg-config@0.29.2
> spdlog@1.9.2
> tclap@1.2.2
> ```
> 
> but Argobot's last commit was back in February so I don't expect to see any change. Time for an Argobots issue or are they aware already?
> 
> from a summit interactive node:
> 
> ```
> $ ../tests/sweep-abt-eventual-bench.sh jsrun -r1 -n1 -a1 perf-regression/abt-eventual-bench
> ```
> 
> ![sweep-abt-eventual-bench 523050 dat cpu](https://user-images.githubusercontent.com/2278417/194925663-7d6e448b-0c4a-403e-bb17-74fba4a021da.png)

Oh right, Ok.  Sorry, I got issues mixed up on that.

I don't think there is any Argobots follow up.  The CPU usage is high, but after thinking on it some more I believe this is a margo behavior.  If all of the threads are running in the same pool, Margo may not be allowing Mercury to sleep during the progress calls.  The CPU utilization is likely lower if we init with a dedicated progress thread.

Thanks for getting this wrapped up @roblatham00 !