@roblatham00 2022-10-10
Kind of a mess.. does this "google perf tool" output confirm your hypothesis?  Lots of time spent in `__margo_hg_progress_fn` and its children, particularly  `NA_Progress` 

[abt-eventual-profile.pdf](https://github.com/mochi-hpc-experiments/mochi-tests/files/9748713/abt-eventual-profile.pdf)