@roblatham00 2022-10-10
One last kick on this dead horse:  modified benchmark to init margo with one dedicated progress thread:

     margo_init("na+sm://", MARGO_CLIENT_MODE, 1, 1);

and the end result is what you'd hope to see:  just about all samples land on futex_wait:

[abt-eventual-progress-profile.pdf](https://github.com/mochi-hpc-experiments/mochi-tests/files/9749152/abt-eventual-progress-profile.pdf)