@roblatham00 2022-10-10
benchmark reports low cpu usage too (meant to include that)

```
#<num_ults>     <test_duration_sec>     <interval_msec> <ops/s> <cpu_time_sec>
100     5       1       1081.000000     0.120891
```